Ensure execution of script in auto-loaded posts
HTML <script> tags don't get executed in HTML that is added as text. The changes in this commit clone any script tags in auto-loaded posts and replace the originals with them.

This also means that our approach to adding post-status to posts can be a lot simpler, but we'll polish that at a later point in time. Note that the javascript for post-status now checks that the document is still loading; using document.write() after the document has finished loading will overwrite the document in its entirety.

Fixes #43